% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cal_Phat.R
\name{Cal_Phat}
\alias{Cal_Phat}
\title{Cal_Phat:calculate the confidence index}
\usage{
Cal_Phat(
  x,
  y,
  A,
  XS = NULL,
  method = c("cos", "frac"),
  frac.par = 0.5,
  sim.par = 0.2
)
}
\arguments{
\item{x:}{n by P feature matrix}

\item{y:}{outcome or residual vector with sizse n}

\item{A:}{treatment, takes value -1 and +1 with size n}

\item{XS:}{names of feature should be used for similarity calculation, if NULL then use all}

\item{method:}{either cosine or fraction similarity, default is cosine "cos"}

\item{frac.par:}{the parameter needed for fraction similarity calculation}

\item{sim.par:}{hyper-parameter governing the definition of neighborhood. It is the minimal similairty quantile for neighborhood}
}
\description{
This is the first step in RICWL.
It calcualtes the confidence index based on the sum of similarity weighted indicator functions.
}
\examples{
n = 200
p = 5
H = matrix(rnorm(n*p), nrow=n)
A = rbinom(n=n, size=1, prob=0.5)*2 - 1
R2 = 4*H[,1]^2 + A*(H[,2]^2 + H[,4]^2 < 1.5)
Cal_Phat(x=H, y=R2, A=A)
}
