% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSVM.R
\name{predict.linearcl}
\alias{predict.linearcl}
\title{predict.linearcl: predict labels for a given new instance when model is trained using linear kernel}
\usage{
\method{predict}{linearcl}(object, x, ...)
}
\arguments{
\item{object:}{wsvm output under linear kernel}

\item{x:}{n by p feature matrix of new subjects}
}
\value{
output label with length n
predict.linearcl()
}
\description{
predict.linearcl: predict labels for a given new instance when model is trained using linear kernel
}
