% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSVM.R
\name{predict.rbfcl}
\alias{predict.rbfcl}
\title{predict.rbfcl: predict labels for a given new instance when model is trained using linear kernel}
\usage{
\method{predict}{rbfcl}(object, x, ...)
}
\arguments{
\item{object:}{wsvm output underl rbf kernel}

\item{x:}{n by p feature matrix of new subjects}
}
\value{
output label with length n
}
\description{
predict.rbfcl: predict labels for a given new instance when model is trained using linear kernel
}
