% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WSVM.R
\name{wsvm}
\alias{wsvm}
\title{wsvm to solve the subject-weighted SVM}
\usage{
wsvm(X, A, wR, kernel = "linear", sigma = 0.05, C = 1, e = 1e-05)
}
\arguments{
\item{X:}{n by p feature matrix}

\item{A:}{label}

\item{wR:}{instance weight}

\item{kernel:}{propensity score with length n, can be either "rbf" or "linear", default is "linear"}

\item{sigma:}{hyper-parameter in SVM rbf kernel}

\item{C:}{hyper-parameter in SVM rbf kernel}

\item{e:}{least tolerated error}
}
\value{
object in rbfcl or linearcf classs
}
\description{
wsvm to solve the subject-weighted SVM
}
